#Comment Fixes to errors found for Tiers 0 and 1 +  Tier 2: Salary > 3700 added #time 10m

+++ src/main/java/io/msalinas/TaxTier0.java
@@ -10,7 +10,7 @@ public class TaxTier0 extends Payslip {
 		super(first, last, salary, superAnnuation, date);
 		super.setUpperLimit(Integer.valueOf(myResources.getString("tier.0.upperLimit")));
 		super.setLowerLimit(Integer.valueOf(myResources.getString("tier.0.lowerLimit")));
-		super.setTaxPerDollar(Integer.valueOf(myResources.getString("tier.0.tax.dollar")));
+		super.setTaxPerDollar(Double.valueOf(myResources.getString("tier.0.tax.dollar")));
 	}
